+++ CHANGELOG.md
@@ -14,6 +14,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Error capturing for python project fires to editor-api rather than Sentry (#625)
+## Added
+
+-  Mobile project details component
+
+### Added
+
+- Mobile navigation between code and output (#615)
+- Mobie sidebar
+
+### Changed
+
+- Position of run button on mobile (#615)
 
 ## [0.18.1] - 2023-08-03